+++ apps/api/Program.cs
@@ -24,6 +24,7 @@ using Mohandseto.Api.Application.AdminInventory;
 using Mohandseto.Api.Application.AdminProcurement;
 using Mohandseto.Api.Application.AdminCrm;
 using Mohandseto.Api.Application.AdminContracts;
+using Mohandseto.Api.Application.AdminPrinting;
 using Mohandseto.Api.Infrastructure;
 using Serilog;
 
@@ -97,6 +98,7 @@ builder.Services.AddScoped<AdminInventoryService>();
 builder.Services.AddScoped<AdminProcurementService>();
 builder.Services.AddScoped<AdminCrmService>();
 builder.Services.AddScoped<AdminContractService>();
+builder.Services.AddScoped<AdminPrintingService>();
 builder.Services.AddHostedService<ContentDispatchWorker>();
 builder.Services.AddHostedService<ContractPriceRevisionWorker>();
 builder.Services.AddScoped<CustomizationService>();